@@ -1,77 +1,75 @@
-name: just_another_workout_timer
+name: 'just_another_workout_timer'
+version: '1.11.2+20220317'
+environment: 
+  sdk: '>=3.3.0'
 description: Just Another Workout Timer
 
-version: 1.11.2+20220317
-
-environment:
-  sdk: ">=3.0.0"
-
 dependencies:
-  flutter:
+  collection: '^1.15.0'
+  dynamic_color: '^1.7.0'
+  flutter: 
     sdk: flutter
-  flutter_tts: ^3.8.3
-  path_provider: 2.1.1
-  flutter_file_dialog: ^3.0.2
-  pref: 2.7.1
-  fluttertoast: ^8.2.2
-  prefs: ^4.1.0+3
-  url_launcher: ^6.1.12
-  wakelock: 0.6.2
-  soundpool: 2.4.1
-  intl_utils: ^2.8.3
-  json_annotation: 4.8.1
-  flutter_phoenix: 1.1.1
-  flutter_native_splash: 2.3.0
-  scrollable_positioned_list: 0.3.8
-  flutter_localizations:
+  flutter_file_dialog: '^3.0.2'
+  flutter_localizations: 
     sdk: flutter
-  uuid: 3.0.7
-#  flutter_nord_theme: 2.1.0
-  share_plus: ^6.3.4
-  provider: 6.0.5
-  dynamic_color: ^1.6.7
-  numberpicker: 2.1.2
+  flutter_native_splash: '^2.4.0'
+  flutter_phoenix: '^1.1.1'
+  flutter_tts: '^4.0.2'
+  fluttertoast: '^8.2.5'
+  intl_utils: '^2.8.7'
+  json_annotation: '^4.9.0'
+  numberpicker: '^2.1.2'
+  path: '^1.8.0'
+  path_provider: '^2.1.3'
+  pref: '^2.8.0'
+  prefs: '^4.1.0+3'
+  provider: '^6.1.2'
+  pubspec_dependency_sorter: '^1.0.5'
+  scrollable_positioned_list: '^0.3.8'
+  share_plus: '^9.0.0'
+  soundpool: '^2.4.1'
+  url_launcher: '^6.2.6'
+  uuid: '^4.4.0'
+  wakelock_plus: '^1.2.5'
 
 dev_dependencies:
-  flutter_launcher_icons: 0.13.1
-  flutter_oss_licenses: 2.0.1
-  json_serializable: ^6.7.1
-  build_runner: ^2.4.6
-  flutter_lints: ^2.0.2
-
-#dependency_overrides:
+  build_runner: '^2.4.10'
+  flutter_launcher_icons: '^0.13.1'
+  flutter_lints: '^4.0.0'
+  flutter_oss_licenses: '^3.0.2'
+  json_serializable: '^6.8.0'
 
 flutter:
   uses-material-design: true
   generate: false
-  assets:
+  assets: 
     - LICENSE
-    - assets/beep_low.wav
-    - assets/beep_high.wav
+    - 'assets/beep_low.wav'
+    - 'assets/beep_high.wav'
     - assets/tick.wav
 
 flutter_icons:
-  android: "launcher_icon"
+  android: 'launcher_icon'
   ios: false
-  image_path: "assets/ic_launcher.png"
-  adaptive_icon_foreground: "assets/ic_launcher_foreground.png"
-  adaptive_icon_background: "#2E3440"
+  image_path: 'assets/ic_launcher.png'
+  adaptive_icon_foreground: 'assets/ic_launcher_foreground.png'
+  adaptive_icon_background: '#2E3440'
 
 flutter_intl:
   enabled: true
 
 flutter_native_splash:
-  color: "#D8DEE9"
-  color_dark: "#2E3440"
-  image: "assets/ic_launcher_foreground.png"
+  color: '#D8DEE9'
+  color_dark: '#2E3440'
+  image: 'assets/ic_launcher_foreground.png'
   fullscreen: true
   android_gravity: center
   android: true
   ios: false
   web: false
-  android12:
-    image: "assets/ic_launcher_foreground.png"
-    color: "#D8DEE9"
-    icon_background_color: "#D8DEE9"
-    color_dark: "#2E3440"
-    icon_background_color_dark: "#2E3440"
+  android12: 
+    image: 'assets/ic_launcher_foreground.png'
+    color: '#D8DEE9'
+    icon_background_color: '#D8DEE9'
+    color_dark: '#2E3440'
+    icon_background_color_dark: '#2E3440'
