@@ -15,7 +15,7 @@ import 'oss_license_page.dart';
 
 /// change some settings of the app and display licenses
 class SettingsPage extends StatefulWidget {
-  const SettingsPage({Key? key}) : super(key: key);
+  const SettingsPage({super.key});
 
   @override
   SettingsPageState createState() => SettingsPageState();
