@@ -9,14 +9,13 @@ import 'utils.dart';
 // ignore: must_be_immutable
 class NumberStepper extends StatefulWidget {
   NumberStepper(
-      {Key? key,
+      {super.key,
       required this.lowerLimit,
       required this.upperLimit,
       required this.value,
       required this.valueChanged,
       required this.formatNumber,
-      required this.largeSteps})
-      : super(key: key);
+      required this.largeSteps});
 
   final int lowerLimit;
   final int upperLimit;
